# src/main/resources/static/api-docs-swagger.yaml
openapi: 3.0.2

info:
  title: Sample API
  description: Sample API without goal yeat
  version: 1.0.0
paths:
  /api/v1/categories/{id}:
    get:
      tags:
        - Categories
      summary: Finds category by id
      description: One id is informed to get a category
      operationId: getCategoryById
      parameters:
        - name: id
          in: path
          description: Id that need to be informed
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryDto"
              example:
                id: 1
                description: "Category 1"
        400:
          description: Invalid id supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponse"
        404:
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponse"
        405:
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponse"
    delete:
      tags:
        - Categories
      summary: Deletes category by id
      description: One id is informed to delete a category
      operationId: deleteCategoryById
      parameters:
        - name: id
          in: path
          description: Id that need to be informed
          required: true
          schema:
            type: integer
      responses:
        204:
          description: Successful operation
        400:
          description: Invalid id supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponse"
        404:
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponse"
        405:
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultResponse"
  /api/v1/categories:
    get:
      tags:
        - Categories
      summary: Get all categories
      description: Gets all categories
      operationId: getAllCategories
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items: {
                  $ref: "#/components/schemas/CategoryDto"
                }
    post:
      tags:
        - Categories
      summary: Create a category
      description: Create a category
      operationId: createCategory
      responses:
        201:
          description: Category created successfully
components:
  schemas:
    DefaultResponse:
      type: object
      properties:
        timestamp:
          type: string
          description: timestamp of the response
          example: "2020-05-30T21:09:11.897+0000"
        status:
          type: integer
          description: http status code
          example: 4xx
        error:
          type: string
          description: Description of the result
          example: "SOME_ERROR"
        message:
          type: string
          description: Message of the response
          example: "Some error"
        path:
            type: string
            description: Path called
            example: "/api/v1/categories/"
    CategoryDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        description:
          type: string
          description: Category description
          example: "Category description"